apps/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
aiosqlite>=0.19.0

# PDF processing
PyMuPDF>=1.23.0
pdf2image>=1.17.0
reportlab>=4.1.0
pypdf>=4.0.0

# Image processing for field detection
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# LLM integration
openai>=1.10.0
anthropic>=0.18.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
